--- project/inung-onboarding/src/main/java/icd/onboarding/surveyproject/service/SurveyService.java
@@ -1,4 +1,0 @@
-package icd.onboarding.surveyproject.service;
-
-public class SurveyService {
-}
