+++ project/inung-onboarding/src/main/java/icd/onboarding/surveyproject/survey/service/SurveyService.java
@@ -1,0 +1,4 @@
+package icd.onboarding.surveyproject.survey.service;
+
+public class SurveyService {
+}
