+++ project/inung-onboarding/src/main/java/icd/onboarding/surveyproject/survey/service/domain/Question.java
@@ -1,7 +1,8 @@
-package icd.onboarding.surveyproject.service.domain;
+package icd.onboarding.surveyproject.survey.service.domain;
 
-import icd.onboarding.surveyproject.service.enums.InputType;
+import icd.onboarding.surveyproject.survey.service.enums.InputType;
 import icd.onboarding.surveyproject.service.exception.*;
+import icd.onboarding.surveyproject.survey.service.exception.*;
 import io.micrometer.common.util.StringUtils;
 import lombok.Getter;
