+++ project/inung-onboarding/src/main/java/icd/onboarding/surveyproject/survey/service/exception/InvalidOptionInfoException.java
@@ -1,4 +1,4 @@
-package icd.onboarding.surveyproject.service.exception;
+package icd.onboarding.surveyproject.survey.service.exception;
 
 public class InvalidOptionInfoException extends RuntimeException{
 }
